Completely revamp attach!ing another mechanism.
Previously, it was conceptually unclear what it meant to attach! mechanism2 to a body of mechanism1, since we were attaching the root body (world) of mechanism2 to mechanism1, and the inertia of mechanism2's root body is undefined. The new interpretation of attach!ing mechanism2 to parentBody (part of mechanism1) is replacing the root body of mechanism2 with parentBody. Note that mechanism2 should be unchanged after attach!ing it to mechanism1. This change also cleans up the function signature.

Fixes #63.

Also made RigidBody and Joint types instead of immutables, among other reasons to help with #58.