Improve URDF parsing for root rigid bodies.
Previously, rigid bodies defined in the URDF without a parent joint were made root rigid bodies of the constructed Mechanism. This change constructs a dedicated root rigid body with undefined inertia, and attaches these bodies to the root with a fixed joint.

Also added functionality to change the type of a Joint after construction of the Mechanism, reenabling the creation of floating base versions of a mechanism loaded from a URDF.