Loop-capable dynamics_solve!
Replace joint_accelerations! with dynamics_solve!, based on Featherstone
p.151. Change DynamicsResult so that it holds data necessary for solving Lagrange
multipliers and joint accelerations for Mechanisms with loops as well.
Change mass matrix to be lower triangular because BLAS/LAPACK functions
appear to be a lot faster that way.