Add maximal_coordinates, test dynamics! with loops
Adds maximal_coordinates, which takes a mechanism and turns it into a
height-one tree with floating joints at the tree edges, while the joints
of the input mechanism become non-tree edges in the
maximal-coordinates-mechanism.

Also fixes a bug in the constraint bias computation.

Test asserts that doing dynamics for a random Mechanism and for a
maximal-coordinates version of that Mechanism results in the same
spatial accelerations.