Stop using canonicalize_graph! in MechanismState ctor (and write_urdf).
Using canocalize_graph! in the MechanismState constructor was necessary
because removing a tree joint can make it so that the indices of the
tree joints are no longer ordered and contiguous. However, it made it so
that constructing MechanismStates was not thread-safe. In addition, it
was perhaps an unexpected (although mostly harmless) behavior to have
the Mechanism be modified in any way upon MechanismState construction or
in write_urdf.

Instead, canonicalize the graph in remove_joint! and
remove_fixed_tree_joints!.