Remove `base` and `body` fields of spatial types.
`GeometricJacobian`, `Twist`, and `SpatialAcceleration` currently have
`body` and `base` frames in addition to the frame in which the
quantities are expressed. This was based on Duindam's definition of a
twist.

However, there are several issues with having the body and base frames
be a part of the spatial quantities.

First, these frames were never all that useful. While matching
the `frame` fields of spatial quantities before operations is a
necessary condition for the operation to make sense, checks involving
the base and body frames (e.g. when adding twists together) were neither
necessary nor sufficient, and often required manually working around the
frame checking mechanism.

Second, the additional frames broke symmetry with the force-space side
of things (`Momentum`, `Wrench`, `MomentumMatrix`, and `WrenchMatrix`),
which lead to reduced opportunities for code deduplication and was kind
of a wart anyway. The other option, adding the additional frames on the
force-space side as well, seems not so great, since frame checks on the
force side are even harder to get right (it would involve keeping a
total balance of action-reaction pairs, so 'local' action/reaction frame
information associated with a single wrench (which may be the sum of
multiple wrenches acting from different bodies) is not rich enough.

Third, I have plans to extract out the frame for all spatial quantities
into a separate type, so that it's also possible to have 'naked'
twists/wrenches/etc. without frame annotations for improved performance
where appropriate (due to additional SIMD opportunities and just less
overhead), as well as to centralize the frame-checking logic. Making
the spatial types more uniform is a step in that direction.

Note that this PR by itself doesn't significantly improve performance
yet.