Change multi-threading behavior of caches (AbstractTypeDicts).
Fix #548.

This is a breaking change, although I doubt that a lot of people have
been relying on this behavior up to this point.

Previously, `AbstractTypeDict` subtypes such as `StateCache`,
`DynamicsResultCache`, and `SegmentedVectorCache` would use the thread
ID for the thread in which the value (e.g. `MechanismState`) was created as part
of the key. This was not needed for thread-safety (a misunderstanding on
my part), and led to some unexpected behavior (#548) as it wasn't
documented properly. I also now consider it to be a poor separation of
concerns: `AbstractTypeDict` should just be `Dict`-like, and `Dict`
doesn't do anything special in the presence of multiple threads.

See the following comments for more info on the past behavior and the
change:

* https://github.com/JuliaRobotics/RigidBodyDynamics.jl/issues/548#issuecomment-491300304
* https://github.com/JuliaRobotics/RigidBodyDynamics.jl/issues/548#issuecomment-491347036